Initial commit.
Script works for the basic files that I had access to on my machine.
Things that look like [Content] Name of Show - 1 [Hash] [Quality
information].mkv will work as will things of the form [Content] Name of
Show 1 - Episode Subtitle [Hash] [Quality].mkv. Added a function that will
save the current names of the files in the folder the Rename-Files method
is executed on just to persist that information. A reversion method will
be made in the future to undo any "damage" done by the renamining script.