New method - Creates symbolic links to files Create-RenamedSymbolicLinksToFiles
New method created called Create-RenamedSymbolicLinksToFiles was created.
This method performs work similar to Rename-Files, but creates a folder of
Plex friendly symbolic links in the same directory that the "renaming"
process works through. This is in case you wish to leave the names of the
actual files as they are (to easily preserve information) while still
getting a Plex friendly file. Feel free to move the symbolic link folder
wherever is convenient for you.

TODO: Make tests, create reversion method to undo what Rename-Files does.